+++ continuum_deployer/resources/deployment.py
@@ -8,10 +8,10 @@ class DeploymentEntity:
     # name of the deployment
     name: str = field(default=None)
     # amount of memory the deployment requires to run
-    memory: float = field(default=0)
+    memory: int = field(default=0)
     # amount of memory the deployment is allowed to use at max
     # TODO - not implemented
-    memory_limit: float = field(default=0)
+    memory_limit: int = field(default=0)
     # amount of cpu resources the deployment requires to run
     # 1=1 cpu core | 0.5=0.5 cpu core
     cpu: float = field(default=0)
